Add files via upload
Detection of mutation. The occurrence of mutational event at the gene level is detected by the altercation it brings 
about in the phenotypic expression of one or more traits of the concerned organism. 
Therefore, the efficiency of detection of mutations will depend largely on the availability of 
techniques for an easy and rapid scoring of the mutant phenotypes in very large populations. 
Scoring of some types of mutations in certain organisms is relatively easy. For example, 
mutations for antibiotic resistance in bacteria are simply detected by plating the bacterial cells 
on a medium containing a lethal concentration of the concerned antibiotic (selective 
medium); the colonies that develop on such a medium will be produced by cells resistant to 
the antibiotic. The medium lacking the antibiotic is called the non selective medium. 
 Number of colonies on the selective medium 
Frequency of mutant cells (%) = -------------------------------------------------- X 100 
 Number of colonies on the non selective medium 
Detection of morphological mutations in eukaryotes requires examination of each individual 
of the population for the mutant phenotype; this is not only tedious requiring time, but is also 
a source of errors in the data. Therefore, elaborate procedures for mutation detection have 
been developed in some eukaryotes . eg; Drosophila, maize etc. These procedures employ 
specific markers to facilitate the identification of chromosomes from the treated or irradiated 
individuals. Clearly these techniques detect only germinal mutations. In drosophila, several 
special genetic stocks have been constructed for the detection of lethal and visible mutations 
in X chromosomes and in autosomes; the two genetic stocks most commonly used for 
mutation detection in X chromosome are (1) CIB and (2) Attached X stocks.